requirements.txt
streamlit==1.31.0
tensorflow==2.15.0
numpy==1.24.3
pillow==10.2.0
opencv-python==4.9.0.80
plotly==5.18.0
pandas==2.1.4
